Rough out the basic type system.
This is basically just transcribing the generated gRPC types into types
that we control. I still need to do the conversion for them.

I changed GetSchema to GetProviderSchema, Stop to StopProvider, and
Configure to ConfigureProvider, as all other RPC calls name the type of
resource they're operating on. This may also provide a little forward
compatibility if we introduce other plugin types to the protocol, e.g.
backends.

I'm still not sure how to represent state. In the gRPC generated types,
it's just a *DynamicValue, which means "msgpack-encoded whatever that
gets a second round of decoding", usually. So I need to figure out what
it gets decoded into and replicate _that_, instead.

I'm going back and forth on the Private []bytes. The protocol just
requires you to echo them back, I believe, which we can do for users
without them needing to deal with these weird opaque fields that they
need to remember to return every time. But "minimal possible
abstraction" has me questioning whether it's our place to do that or
not.

Representing cty is also something I need to figure out. Once again, its
protocol definition is just a *DynamicValue, which only tells us that it
will be decoded _again_ by the application, not what it's decoded into,
which is what we want to surface. Should be easy enough to find, but it
wasn't in front of my face in the protobuf definition, so I didn't
include it in this pass.

I have no idea why SchemaAttribute.Type is a []byte in the protobuf, and
want someone to explain that to me and why it isn't, say, a string, so I
can include it as a comment in the code, because that's bizarre.

I made all the code happen in tfprotov5 instead of the root package or a
tfproto package because I want to be able to have multiple major
versions of the protocol in a single module version, so we don't have to
do weird long-running branches to support v5 and v6 simultaneously if we
ever have a v6.

I used a system of composition to do some weird type stuff here. The
ResourceServer and DataSourceServer interfaces handle all the RPC
calls for a resource or a data source, respectively. This allows us to
have helpers that just send resource calls to a resource type, so users
don't _need_ to branch all their resources out of ApplyResourceChange or
whatever. Then a ProviderServer embeds those interfaces, relying on all
their methods, and adds the provider-specific ones.

I also used this approach to create a BaseResourceServer,
BaseDataSourceServer, and BaseProviderServer that handle their unique
methods (BaseProviderServer anonymously embeds BaseResourceServer and
BaseDataSourceServer) and return a "not implemented" error message. This
is an idea I'm playing with for forward-compatible interfaces; a struct
can anonymously embed these types, and it will automatically implement
the interface, even if the interface expands underneath it. I'm not sure
if this is better than a compile-time warning, or if we can provide
reasonable default behavior for some or all of these methods, but it
seemed like an idea worth trying.