Correct some of our state and cty code.
Some state was incorrectly typed as RawState, when it's really a
DynamicValue--meaning cty.

Instead of exposing cty, we're going to call it "tftypes" or "Terraform
types" in the SDK, as cty is an implementation detail. Move all our cty
stub stuff into a tftypes package, rename it accordingly, and add some
basic helper structure to it to begin roughing out what that may look
like eventually.