Flesh out tftypes package a bit more.
Our type system is starting to take shape. It leans a little more
towards capturing the guts of cty's implementation details and smoothing
over them with semantics using the `RawValue.Unmarshal` method.

There's a bunch of ambiguity in cty types and how we get type
information. It's not captured in cty when encoding with msgpack whether
something is a naked set, tuple, or list, or whether it is type
information then a value. Similarly, JSON doesn't capture whether
something is a map, object, or type information and value. The decoder
needs to know this information or push that problem on to the consumer.

Our implementation is choosing to capture the raw ambiguity in the
RawValue type, and then paper it over in Unmarshal based on the type
that the user provides to us.