Fix build errors.
Types are now strings, because why make life harder than it needs to be?
Strings and Sets clashed on accident when it was runes, and that's a
silly and unnecessary bug to have to save bytes that probably don't
matter.

Fix how we're assigning results to dst so that it'll actually compile.
Whoops.