Add resource and data source routers.
Add a ResourceServer and DataSourceServer implementation that will route
requests based on the request's TypeName to other ResourceServer and
DataSourceServer implementations.

This can be used to delegate handling of each resource's and data
source's requests to functions specific to those resources and data
sources, so the logic isn't all mixed together in one big function.