Add JSON and msgpack unmarshaling.
We unmarshal from []byte values into a tftypes.RawValue, using the basic
empty interface decoding provided to both msgpack and json. This allows
us to present a single, unified abstraction to provider developers
instead of making them worry about different encodings.

There is a problem with this approach, however: we don't specify what
the behavior of that empty interface decoding _is_. Meaning it could
change out from under us, and more importantly, msgpack and JSON don't
necessarily need to agree on how to represent something. Neither of
those properties is ideal for a provider developer, so we should smooth
them over by manually specifying the behavior, perhaps with a custom
decoder (which is what cty does) so we can provide a clear mapping of
cty types to Go types.

Developers should still call the RawValue.Unmarshal method before
dealing with any of these values. We need a big scary GoDoc to that
effect still, but the idea is that providers can type convert from one
Go type to another, instead of dealing with the implementation details
of encodings.