Add some handling for complex types.
Complex types, or values that come back with type information inline,
require some specialised handling and parsing. This gets us most of the
way there, with several TODOs remaining to polish it up a bit before
it's actually ready.

We're going to need extensive unit tests around this logic, because it's
probably the most dangerous area of code we have right now in terms of
complexity.