Complicate types.
cty tracks not only the top-level type, but the types of all attributes
and elements in its type information. To properly surface cty, then, we
need to track all that type information as part of the type, too.

Obviously, our old string alias Type type wasn't going to be able to do
that, so we needed to complicate things a bit.

Turning Type into a struct and giving it properties for AttributeType,
AttributeTypes, ElementType, and ElementTypes worked until we got to
complex types with those properties filled in. Each instance would be a
new instance of the Type type, which wasn't ideal; we could no longer
share a single Type variable for all lists the way we used to.

So Type became an interface, with String, Number, and Bool all using a
single `primitive` type, which implemented the interface and is largely
the same code as our old Type type. Set, Object, Map, List, and Tuple
all got their own implementations.

Our unmarshaling code obviously needed some updating, because we had to
populate those new type fields. This led to parsing the type information
provided by cty _first_, and _then_ trying to associate it with values.
This lets us populate the type of the top-level type in one shot, and
also lets us just refer back to the pre-parsed type information instead
of trying to parse it at every level while parsing the values. This led
to a lot less code and it's much easier to follow.

The default types based on Go types also needed to change; List can't be
the default, because there may be items in the list of different types,
which List doesn't support. Same with Map. So we needed to change those
to Object and Tuple, respectively, which can support polymorphic items.