Make all Numbers *big.Floats by default.
When handling types that came from JSON or msgpack, anything that's
supposed to be a number we try to coerce into being a *big.Float, which
is the only lossless type we can standardise all Numbers on. This
prevents developers from needing to worry about getting Numbers in
different intermediary types based on what value was provided. Unmarshal
will give developers the type in whatever convenient format they ask
for, but will always throw an error in the face of a loss of specificity
(except for float32/float64, in which case it's impossible to know if
accuracy was lost).