Complicate type system.
The old type system was simple, but insufficient. We erased some
necessary complexity by ignoring unknown values. You could represent
them as tftypes.RawValues, but the intended usage of a tftypes.RawValue
is to call its Unmarshal method as close to the boundary of your
provider as possible. Unfortunately, in that situation, we never really
thought through what would happen with unknown values, which can't be
represented in a strongly-typed Go type, they would need to be an
interface or struct of some sort. So that approach wouldn't work, in
practice.

This new approach is to expose tfprotov5.DynamicValues from the protocol
methods, and they have an Unmarshal method that gives you a
tftypes.Value. The DynamicValue is just the msgpack/json bytes, the
Unmarshal method takes the tftypes.Type you want them decoded into, and
returns a tftypes.Value of that type if possible.

tftypes.Value gained some new helper methods, most notably As(), which
lets you convert from the canonical types that JSON and msgpack will
store in the Value into a more friendly type, if that conversion is
possible. So far, the only conversions that are possible is calling the
UnmarshalTerraform5Type method if the destination has one. A future
improvement would be to build in some default conversions for the
builtin Go types and maybe some collection types for tftypes.Values,
like []tftypes.Value or map[string]tftypes.Value. This approach gives us
a more explicit interface that's decoupled from the type we store in the
Value, and it allows us to massage things if we change details or
introduce new types, so there's not a backwards compatibility concern.

Finally, we added a tftypes.Path type, which is used to keep track of
pathing information into values so we can return more useful, specific
errors. It was integrated into the JSON unmarshaling and msgpack
marshaling, but probably could stand to be integrated into the msgpack
unmarshaling and to have an easy, handy conversion into a
tfprotov5.AttributePath. That may require some minor reworking.