Let there be AttributePaths!
Finally connect the dots on AttributePaths. We move the type into
tftypes, so it can be used throughout the tfprotov5 and the tftypes
package without causing import cycles (tfprotov5 imports tftypes
already). We get rid of our old tftypes.Path type, and just stick with
one type of paths. And finally, msgpack unmarshaling gets upgraded to
use the new tftypes.AttributePaths for enhanced errors.

This does not add any helpers for extracting the AttributePath from an
attributePathError, but we should do that, so we can bubble up
attributePathErrors and get the AttributePath they're referring to for
diagnostic purposes.

Another day.