Support the legacy type system.
Unfortunately, as SDKv2 needs to be built on this, so we need to export
the flag in ApplyResourceChange and PlanResourceChange that tells
Terraform to permit SDKv2's legacy shenanigans. But nobody else should
ever use ever for any reason at all it has that one purpose and that one
purpose only. So we gave it a scary name and a scary GoDoc message and
basically tried to warn people off it as much as possible.