Flesh out the As method.
The As method on tftypes.Values was essentially useless, because As is
the only way to get to the underlying Go types beneath a Value, but As
takes a Value as its argument, and As is the only thing you can call on
a Value. So there was nothing you could actually do inside your
Unmarshaler method because As only supported Unmarshalers. So As would
call your Unmarshaler which could only call As which could only call
your Unmarshaler and so on and so forth.

This implements some basic Go type support for the canonical Go
representations of all the tftypes Types, so we can get to them in our
Unmarshaler methods or just use them with As directly.