Fix bug in msgpack invalid value error.
Our invalid value type error when marshaling to msgpack is supposed to
tell people what type they passed in. Unfortunately, we're telling them
the type of the Value they passed in, which is always tftypes.Value, not
the *contents* of that Value, which is the actual type they should know
about. This leads to confusing error messages that aren't actually
helpful.

Switching to pass in the content of the tftypes.Value gives a much
better error message.