Actually handle null values.
When calling `tftypes.Value.As` actually handle null values. If casting
to a pointer to a pointer, null values set the pointer to nil. If casting
to just a pointer, null values set the variable to its empty value,
because we can't safely assume we can store nil; the practice of calling
`Value.As(&dst)` is too common, in which case `dst` wouldn't retain the
nil the way users would expect. So we require a pointer to a pointer to
be able to surface the nil value.