Add serve helper function.
Add a helper that will stand up the go-plugin server for us.

I made it require a name for the provider because I think having that
information will be useful to us down the road for things like logging.

It takes a `func() tfprotov5.ProviderServer`, which it will use to
instantiate a tfprotov5.ProviderServer, which it will then serve.
Factories only this time.

Finally, I made it use the variadic options pattern so we can add more
options later and offer a clearer indication of what's required when. I
added two sample implementations of providing your own logger to
go-plugin and of using the test mode of go-plugin that allows for
reattaching to running provider processes.