Disable golint for toproto.
I'm not happy about this, but golint got mad about the underscores in
our function names, so we need to either change the names, stop using
golint, or disable golint in those files. We don't want to change the
names, because they intentionally match the names generated by gRPC that
they're translating to. We don't want to stop using golint, because it's
useful. That leaves us with disabling golint for those files, which we
do by pretending they're generated even though they are not. This does
lose us all other golint checking for those files, which makes me very
sad, but here we are.