Rename our UnmarshalTerraform5Type interface.
The names didn't make a whole lot of sense. "Unmarshal" evokes JSON or
other encoding schemes, which isn't really what we're doing here. Also,
we're not working with Types, we're working wtih Values.

By using FromTerraform5Value, we can mirror it with ToTerraform5Value,
and the name's are clearer about what the results will be and their
usage.