Move msgpack tests to tfprotov5, implement parsing.
Move our msgpack encoding tests to the tfprotov5 package, which led to
some unfortunate exports needing to be added to the tftypes package,
mostly the ValueComparer function.

But this lets us quickly and easily reuse those tests in the tfprotov5
package to test that our parsing of msgpack values is correct. We now
mimic the cty tests incredibly closely.