Implement unit tests (#32)
* Test marshaling msgpack.

Test that the bytes produced from msgpack marshaling match the bytes cty
produces for equivalent msgpack marshaling. This hardcodes the bytes so
we don't have a dependency on cty.

* Fix dynamic tests.

* Make tests parallel.

* Test behavior of Value.As.

Add tests that assert we get the results we expect for given inputs of
the tftypes.Value.As method.

* Add tests for encoding and decoding types with JSON.

Add a test that will decode JSON from a hard-coded string and assert
that the type that was parsed is correct, then will encode the type and
ensure that the generated JSON is correct.

* Test Value.IsKnown() and Value.IsFullyKnown()

* Test WalkAttributePath.

* Move msgpack tests to tfprotov5, implement parsing.

Move our msgpack encoding tests to the tfprotov5 package, which led to
some unfortunate exports needing to be added to the tftypes package,
mostly the ValueComparer function.

But this lets us quickly and easily reuse those tests in the tfprotov5
package to test that our parsing of msgpack values is correct. We now
mimic the cty tests incredibly closely.

* Add tests for parsing JSON from DynamicValues.