Expose Types of tftypes.Values.
Add a method to expose the tftypes.Type associated with a tftypes.Value.
This is necessary because tftypes.Value.Is is not always able to
reconstruct a type from comparison alone; for example, reconstructing
what types a map, object, list, set, or tuple uses when its
tftypes.Value is the unknown value is impossible.

But we can make it so people can't directly compare them with ==, which
we don't want them doing (we want them to use Is instead) by embedding a
`_ []struct{}` in each type, which cannot be compared and prevents the
types from being compared, raising a compile time error.

This does mean our primitive types are now structs, and therefore cannot
be constants, which is sad. In theory, someone could overwrite them and
change the output, but the amount of shenanigans one can get up to with
this is pretty limited. You could rename a type to another type:

```go
tftypes.DynamicPseudoType = tftypes.String
```

but you can't construct a new type out of whole cloth, because the
`primitive` type isn't exported and so you can't create new values for
it. So people could wreak some havoc here if they were malicious, but
this doesn't _seem_ to expose us to any new vectors of user error.