Make AttributePaths pointers, fix AttributePathErrors.
Make AttributePaths into pointers everywhere, so they can be
ergonomically used with diagnostics. nil pointers are considered
equivalent to AttributePaths with no steps.

Because modifying steps on the AttributePath without defensively copying
first is now A Major Headache and makes everything awful, I made them
private. This necessitated the creation of a NewAttributePathWithSteps
so our internal code for translating between protocol and tftypes could
create the AttributePath all in one go without a bunch of unnecessary
code. I also added NewAttributePath to avoid littering
`&AttributePath{}` around our code and provider code everywhere.

This also necessitated adding a `Steps` method to AttributePaths, so
provider devs could get at the steps in a read-only fashion. `Steps` is
safe to call even if the AttributePath it's called on is nil, it will
just return an empty slice.

AttributePathErrors became exported, so provider devs can detect them,
and the Path was exported so provider devs could get at it. It also
gained its own Error method, prefixing the error message it's wrapping
with the AttributePath it's associating that error with. It also got an
Unwrap method, supporting `errors.Is` and `errors.As`.

The rest of this is just updating our code to use these new types and
methods, as AttributePaths are used... well, basically everywhere.

Fixes #61.
Fixes #10.
Fixes #34.