Implement OptionalAttributes for Objects.
This is another massive PR, sorry. I tried to separate out some of the
changes, but the code kept overlapping, unfortunately.

All our Types got an Equal method, so they can be compared using go-cmp.
This was necessary to make the tests pass. As part of this, the Type
interface was refactored to include an equals() private method, which
powers the Equal and Is methods on each type.

I also fixed some periods in error messages to make golint happy.

Of course, with these new methods, there was enough logic to add tests
to each Type, so I added some tests that exercised the equality checks.

I also added, which was the main thing I was trying to do here, an
OptionalAttributes property to Objects. These are attributes that _can_
be set on an object, but are not _required_ to be.

Because OptionalAttributes made the NewValue logic more complicated for
Objects, I wrote some tests for creating Objects with NewValue to test
the logic.

Finally, the new Object comparison logic was panicking during a
tfprotov5 / tfprotov6 test, because the Object had a nil AttributeTypes
map instead of an empty AttributeTypes map. I fixed it to use an empty
AttributeTypes map. In theory, I don't think we should ever see a nil
AttributeTypes map.