Refactor JSON and MsgPack parsing code. (#77)
We previously had JSON and MsgPack parsing code implemented as private functions in each protocol implementation. This was fine, as far as it went, but it didn't let us share our implementations across protocol implementations. Given this is really more about the type system than the protocol, it makes more sense to keep it with the tftypes package and have a single implementation. This has the nice side-effect of moving it into the same package its tests are implemented in, which lets us get coverage information out of them, though there are other ways we could've accomplished that.