Fix panic for Value.As(**big.Float). (#85)
*big.Float is a pointer type that will never really appear as anything
except a pointer. So when using code like this:

```go
var foo *big.Float
err := value.As(&foo)
```

the target of `As` is actually a **big.Float. But our code dereferences
it and calls `As` on the target of the pointer, but the target of the
pointer is actually a nil pointer.

This code fixes that by instantiating a new value for the pointer to
point to, setting the pointer to that, and then calling `As` on the new
value.