Don't panic when String() is called on nil Value{} (#86)
Currently, attempting to call val.String() where val := Value{} results
in a panic.

While a Value without a type will not be returned from NewValue except
in case of error, Value{} is likely to appear in consumer code from time
to time, e.g. in tests, as it is the nil value of the struct. Define an
arbitrary string representation for Value{}.

Backwards-compatible because this was previously undefined behaviour.