Add terraform-plugin-log support.
Set up loggers using terraform-plugin-log v0.1.0, and update our
tfprotov5 and tfprotov6 server packages to have trace and error logs,
and to inject useful values in requests opportunistically.

Let providers configure the name for logging purposes and to control the
environment variable that is used for the provider's log level. It
defaults to the shortname of the provider from the registry address
provided, or no environment variable control if the registry address
can't be parsed.

Make it so we can inject a log sink.