Start refactoring type comparisons.
Update our Type interface to include UsableAs and Equal explicitly.

Add placeholders for UsableAs and Equal on all types, so tests can run.

Update String, Number, Bool, and DynamicPseudoType to have UsableAs
implementations, and tests for those implementations. DynamicPseudoType
will panic when UsableAs is called on it, because DynamicPseudoType is a
type constraint, it should never be associated with a Value.

This work will allow us to resolve #83.