tfprotov5+tfprotov6: Remove unnecessary AttributePath_Step error return
Reference: https://github.com/hashicorp/terraform-plugin-go/issues/365

It is not possible for an unexpected attribute path step implementation to reach the protocol conversion logic unless there is a protocol change, which the logic should be updated anyways at the same time and unit tested. This simplifies many of the gRPC server handling methods to not need to handle the currently impossible-to-have errors.